Apply Slip Days

This function applies any slip day policies within the policy file.

Usage

apply_slip_days(gs, policy)

Arguments

gs

A Gradescope dataframe

policy

A grading policy file as an R list

Value

A dataframe of the original Gradescope data with lateness updated according to the slip day policies

Examples

path <- system.file("extdata", "slip-days-policy-demo.yaml", package = "gradebook")
slip_days_policy_demo <- read_policy(path = path)
apply_slip_days(gs = gs_demo, policy = slip_days_policy_demo)
#> # A tibble: 10 × 45
#>    Names               Email     SID Sections `Lab 1` `Lab 1 - Max Points`
#>    <chr>               <chr>   <dbl> <chr>      <dbl>                <dbl>
#>  1 Dr. Kadin           Dr.Ka… 3.03e9 fake-cl…       9                   10
#>  2 Bartley Toy         Bartl… 3.03e9 fake-cl…       5                   10
#>  3 Dr. Dontae          Dr.Do… 3.03e9 fake-cl…      10                   10
#>  4 Narcissus Considine Narci… 3.03e9 fake-cl…      10                   10
#>  5 Verne Wilderman     Verne… 3.03e9 fake-cl…       8                   10
#>  6 Toivo Witting       Toivo… 3.03e9 fake-cl…       9                   10
#>  7 Tegan Crooks        Tegan… 3.03e9 fake-cl…       8                   10
#>  8 Tyrus Cremin-Wisozk Tyrus… 3.03e9 fake-cl…       2                   10
#>  9 Sydni Hansen        Sydni… 3.03e9 fake-cl…       8                   10
#> 10 Mrs. Chelsie        Mrs.C… 3.03e9 fake-cl…       5                   10
#> # ℹ 39 more variables: `Lab 1 - Submission Time` <chr>,
#> #   `Lab 1 - Lateness (H:M:S)` <chr>, `Lab 2` <dbl>,
#> #   `Lab 2 - Max Points` <dbl>, `Lab 2 - Submission Time` <chr>,
#> #   `Lab 2 - Lateness (H:M:S)` <chr>, `Lab 3` <dbl>,
#> #   `Lab 3 - Max Points` <dbl>, `Lab 3 - Submission Time` <chr>,
#> #   `Lab 3 - Lateness (H:M:S)` <chr>, `Lab 4` <dbl>,
#> #   `Lab 4 - Max Points` <dbl>, `Lab 4 - Submission Time` <chr>, …