Reshape policy file from nested to flat

Cycles through the top-level categories of a policy file and unnested all subcategories to create a single level list of all categories and subcategories.

Usage

flatten_policy(policy)

Arguments

policy

A valid policy file stored as a list.

Value

A single level list of all categories and subcategories ordered where all leaves will precede the parent category in the list order.

Examples

# Example
flatten_policy(policy_demo)
#> $coursewide
#> $coursewide$course_name
#> [1] "Demo Course"
#> 
#> $coursewide$description
#> [1] "This is a demo class dataset to demonstrate the app."
#> 
#> 
#> $categories
#> $categories[[1]]
#> $categories[[1]]$name
#> [1] "Labs"
#> 
#> $categories[[1]]$slipdays
#> [1] 0
#> 
#> $categories[[1]]$late_time1
#> [1] "00:00:00"
#> 
#> $categories[[1]]$late_time2
#> [1] "00:00:01"
#> 
#> $categories[[1]]$late_scale1
#> [1] 1
#> 
#> $categories[[1]]$late_scale2
#> [1] 0
#> 
#> $categories[[1]]$after
#> [1] TRUE
#> 
#> $categories[[1]]$weight
#> [1] 55
#> 
#> $categories[[1]]$drops
#> [1] 2
#> 
#> $categories[[1]]$weighted_method
#> [1] "equally"   "by_points"
#> 
#> $categories[[1]]$clobber
#> [1] "None"
#> 
#> $categories[[1]]$assigns
#> [1] "lab_1" "lab_2" "lab_3" "lab_4" "lab_5" "lab_6"
#> 
#> 
#> $categories[[2]]
#> $categories[[2]]$name
#> [1] "Discussions"
#> 
#> $categories[[2]]$slipdays
#> [1] 0
#> 
#> $categories[[2]]$late_time1
#> [1] "00:00:01"
#> 
#> $categories[[2]]$late_time2
#> [1] "00:00:00"
#> 
#> $categories[[2]]$late_scale1
#> [1] 1
#> 
#> $categories[[2]]$late_scale2
#> [1] 0
#> 
#> $categories[[2]]$after
#> [1] TRUE
#> 
#> $categories[[2]]$weight
#> [1] 10
#> 
#> $categories[[2]]$drops
#> [1] 0
#> 
#> $categories[[2]]$weighted_method
#> [1] "equally"   "by_points"
#> 
#> $categories[[2]]$clobber
#> [1] "None"
#> 
#> $categories[[2]]$assigns
#> [1] "discussion_1" "discussion_2"
#> 
#> 
#> $categories[[3]]
#> $categories[[3]]$name
#> [1] "Midterm"
#> 
#> $categories[[3]]$slipdays
#> [1] 0
#> 
#> $categories[[3]]$late_time1
#> [1] "00:00:01"
#> 
#> $categories[[3]]$late_time2
#> [1] "00:00:00"
#> 
#> $categories[[3]]$late_scale1
#> [1] 1
#> 
#> $categories[[3]]$late_scale2
#> [1] 0
#> 
#> $categories[[3]]$after
#> [1] TRUE
#> 
#> $categories[[3]]$weight
#> [1] 15
#> 
#> $categories[[3]]$drops
#> [1] 0
#> 
#> $categories[[3]]$weighted_method
#> [1] "equally"   "by_points"
#> 
#> $categories[[3]]$clobber
#> [1] "None"
#> 
#> $categories[[3]]$assigns
#> [1] "midterm"
#> 
#> 
#> $categories[[4]]
#> $categories[[4]]$name
#> [1] "Final"
#> 
#> $categories[[4]]$slipdays
#> [1] 0
#> 
#> $categories[[4]]$late_time1
#> [1] "00:00:01"
#> 
#> $categories[[4]]$late_time2
#> [1] "00:00:00"
#> 
#> $categories[[4]]$late_scale1
#> [1] 1
#> 
#> $categories[[4]]$late_scale2
#> [1] 0
#> 
#> $categories[[4]]$after
#> [1] TRUE
#> 
#> $categories[[4]]$weight
#> [1] 20
#> 
#> $categories[[4]]$drops
#> [1] 0
#> 
#> $categories[[4]]$weighted_method
#> [1] "equally"   "by_points"
#> 
#> $categories[[4]]$clobber
#> [1] "None"
#> 
#> $categories[[4]]$assigns
#> [1] "final"
#> 
#> 
#> 
#> $cutoff
#> $cutoff$A
#> [1] 90
#> 
#> $cutoff$B
#> [1] 80
#> 
#> $cutoff$C
#> [1] 70
#> 
#> $cutoff$D
#> [1] 60
#> 
#> $cutoff$F
#> [1] 0
#> 
#>