Simple Policy Demo File

A demo policy file that mimics a hypothetical class syllabus

Usage

simple_policy

Format

A list of three nested lists

COURSEWIDE

first nested list with general information about the class

course_name

name of the course

description

any relevant course description, purely for the user

CATEGORIES

second nested list where syllabus is broken down into categories with their respective grading criteria

name

category name

slipdays

number of slip days

late_time1, late_time2

cutoffs for lateness penalty

late_scale1, late_scale2

lateness penalty for respective lateness cutoffs

after

if TRUE, late_scale1 is applied AFTER late_time1 has passed; if FALSE, late_scale1 is applied UNTIL late_scale2 has passed; same for late_scale2 and late_time2

weight

weight of this category in overall grade

drops

the number of lowest-score assignments that are dropped from the grade

weighted_equally

if TRUE, all assignments are weighted by their percentages; if FALSE, all assignments are weighted by their max-point value

clobber

if the clobber-category has a higher grade than this one, the clobber-category's grade replaces the grade for this category

assigns

list of assignment names that falls into this category

CUTOFF

third and final nested list where each cutoff reflects the lower bound cutoff for each letter grade: A,B,C,D,F

...

Examples

simple_policy
#> $coursewide
#> $coursewide$course
#> [1] "Stat 101"
#> 
#> 
#> $categories
#> $categories[[1]]
#> $categories[[1]]$category
#> [1] "Labs"
#> 
#> $categories[[1]]$aggregation
#> [1] "weighted_by_points"
#> 
#> $categories[[1]]$weight
#> [1] 0.3
#> 
#> $categories[[1]]$assignments
#> [1] "Lab 1" "Lab 2" "Lab 3" "Lab 4" "Lab 5" "Lab 6"
#> 
#> 
#> $categories[[2]]
#> $categories[[2]]$category
#> [1] "Discussion"
#> 
#> $categories[[2]]$aggregation
#> [1] "equally_weighted"
#> 
#> $categories[[2]]$weight
#> [1] 0.1
#> 
#> $categories[[2]]$assignments
#> [1] "Discussion 1" "Discussion 2"
#> 
#> 
#> $categories[[3]]
#> $categories[[3]]$category
#> [1] "Midterm"
#> 
#> $categories[[3]]$aggregation
#> [1] "none"
#> 
#> $categories[[3]]$weight
#> [1] 0.3
#> 
#> $categories[[3]]$assignments
#> [1] "Midterm"
#> 
#> 
#> $categories[[4]]
#> $categories[[4]]$category
#> [1] "Final"
#> 
#> $categories[[4]]$aggregation
#> [1] "none"
#> 
#> $categories[[4]]$weight
#> [1] 0.3
#> 
#> $categories[[4]]$assignments
#> [1] "Final"
#> 
#> 
#> 

# yaml:as_yaml(simple_policy)